This works just lovely!
HOWEVER, it is also clear that running it once, twice, three times,
for EACH generator looking for good ones is a PITA.  We'll have to
eventually rearrange this so that there is a "search mode" that runs
a loop through all known generators, identifying the ones that pass at
least at the 1% or higher level.

BTW, I'm now prepared to bet a nickel that the rgb binomial test has a
great deal of sensitivity, since it fails all but literally three or
four of the available RNG's for absurdly short data strings.  As in they
aren't even APPROXIMATELY random...NONE of them.  If one used them to
generate a humble binomial distribution numerically it would be in
significant error.

I do need to alter this test so that I can run it for arbitrary bit
string lengths, but for the moment I'm not going to worry about it.