This is a pretty serious bugfix -- probably need to update the website. Basically, my kstest was simply wrong last night; now it is working, I've also added the Kuiper form of the KS test, and will program Anderson and Darling's version (the one used, apparently, in Diehard) when I get around to it.  However, for tests involving more than a very few p-values in a vector, it shouldn't really matter -- Kuiper KS and the regular KS and Anderson-Davis KS should all GENERALLY generate similar p-distributions -- different perhaps where they don't matter, but very similar at the ends.  The key question is just whether one has a tendency to pass a vector of p-values where the other would consistently fail it.  So far it looks like USUALLY if one fails the other fails.
I think I'm still going to want to do a histogram picture of binned
pvalues and do a Pearson chisq p-test on the result.  This should really
be pretty easy... maybe today, maybe not.

We're getting close to being ready to go BACK and mess with the Ntest
and Xtest stuff.  I think that now that I understand Pearson vs the
alternatives, I can PROBABLY arrange things so that I can use a single
set of common tools to do all the test assessment.

One thing, for example, would be to make each test return a p-value,
period, and put the samples loop in rand_rate_work, to ALWAYS fill a
vector of p-values and ALWAYS do KS tests, confidence interval tests,
and histogram tests.  This would have a number of advantages -- being
able to produce a really pretty, really standardized picture of results
for one.

A second thing that would make this tool relatively interesting to the
mass unwashed would be to put a nice little GUI onto it.  There are two
generic ways to do this.  One is to leave it a command line tool but
REALLY clean up the output result so that it is just a single line
per test with ks test scores for the various forms of test with three
lines of # delimited frame, period.  Then I can make a perl-Gtk app to
call the binary, parse the result, and (e.g.) plot histograms or do
other nifty graphical things.  The other is to use Gtk directly, but
perhaps have the GUI only come up if there is an appropriate command
line flag (or not).

A third thing to work on is clearly going to be splitting the source
into distinct components in distinct directories.  We will need a common
library containing the kstest, chisq, Ntest, Xtest etc code, the input,
the output, etc.  We will need a directory containing extensions to the
GSL random number library, e.g. /dev/random, /dev/urandom, empty, and
shuffled because the one thing that is absolutely true is that we need
to add a shuffling/mixing random number generator, one that permits us
to set up a shuffling list and refill it from a secondary LIST of rng's!

A fourth thing (noted already elsewhere) is to do the simplest of tests
-- apply a KS test to the GSL distribution-specific generators
themselves.  If a "test" is generically generating a known distribution
presuming randomness and then seeing if the result is indeed the
targeted distribution, then EVERY distribution generator in the GSL can
simultaneously be the target of a test for algorithmic purposes AND a
test component for the GSL rng's.

Beyond that, I need to implement spectral tests and tests for
hyperplanes in N dimensions and uniformity tests.  Sigh.  I think that I
DO need to write a grant proposal for this -- I think there is enough
work to justify it.