A bugfix commit.  The sanity check in get_bit() is broken and is commented out -- if I'm going to allocate rand_int[] vectors other than size in length, I cannot test on a global size to see if get_bit() is out of bounds.  This is STILL broken in that there is a risk with no warning, but there is also functionality for the moment (and I have to write a bunch of new bitlevel functions and can rewrite get_bits at the same time).
More important, I found a real bug in Btest where I was initializing
btest->chisq to zero before accumulation but was accumulating in chisq.
Curiously, it worked a lot of the time the old way, but only for certain
rng's.  I may have memory management problem, which isn't surprising
given the slovenliness of the code at this moment...;-)

   rgb