This should/maybe be a serious v 0.5.3 checkin.  We are about to try -v -1.  If it works, it will cycle through all of the working tests (and all the tests are working).  All the tests are now written in a way that they can use sample and kstest_kuiper() to do the validation of the p-values obtained from running a possibly size-variable test on bits or frequencies or runs or whatever.
If this test works, it is off to the website, I'm off to bed, and next
we go back to moving in new diehard tests.  With the magical sliding
bitwindow (which really does seem to work pretty well:-) implementing
at least the n-tuple diehard tests should now be pretty easy, and I
can probably do a more rigorous job than GM did because I don't have
to scrimp on rands.