Checking in several things.  rgb_operm.c is a template for my eventual replacement for operm5.  I have about concluded that operm5 is basically a joke -- it is conceptually incorrect.  The correct statistic should be the overlapping permutations covariance matrix itself.  This test just counts the occurrence of permutations -- the "overlap" part is completely irrelevant in this context.  It returns an aggregate count vector whose mean is 1/120 (obviously, permutations of 5 objects = 5! = 120).  It makes NO DIFFERENCE AT ALL if this vector is evaluated from overlapping samples or from independent samples -- the mean will be the same.  All that MIGHT change is the variance -- since the samples are drawn from an overlapping population, one would expect that their variance would be strictly less than one would expect from truly independent samples.
In a minute I'll verify this with the overlap flag and a couple of runs
with mt, but I now understand why the test fails.  It remains to figure
out how to build a CORRECT overlapping permutations test that uses the
COVARIANCE of overlapping samples as a statistic.  I actually think that
this will be pretty easy, but I do have to also understand why the
definition in the nilpotent markov chains paper doesn't (I think) lead
to the numbers that they publish there.  Or rather, why my MUCH SIMPLER
algorithm doesn't lead to the same numbers they get.