We are working HARD or confirming or denying dieharder_operm5 AND working out a more general rgb_operm.  The convertoperm.pl script once AGAIN confirms that I'm unpacking and repacking (for C) the r and s matrices of diehard_operm correctly.
However, AFAICT, the diehard_operm5 JUST tallies up t[120], the count of
the different permutation indices.  However, all the permutations should
occur with a SYMMETRIC probability -- I verified this directly by doing
the actual integrals, but in retrospect it is pretty obvious from
symmetry alone.  The t-vector therefore has NOTHING TO DO with the
covariance matrix no matter what you do with it, which may be why
diehard_operm5() is a bad test.  FWIW, I've added output statements to
diehard_operm5() that dump the t-vector and verified empirically that
yes, it produces a more or less uniform distribution of bin-probability
1/120 -- one could in fact do a chisq test on this directly and very
likely obtain an actual meaningful test, but that is not at all what is
done.  

Note well that the sampled distribution of t[] will not SIGNIFICANTLY
vary if computed from overlapping samples!  This is simple to see -- it
would be just like computing five distinct samplings of non-overlapping
samples, each of which would be expected to have exactly the same mean
and variance.  Because of the overlap there could be a tiny bit of
correlation between the FLUCTUATIONS in the MEASURED mean values for any
given bin, but the bin mean would have to be unchanged for each
overlapping sampling, and for that matter the bin variance for each
sampling would ALSO have to be the same. 

The only effect of combining the overlapping five samplings (each with
identical mean and variance!)  would therefore be to AT MOST alter the
sampled VARIANCE of the bin distribution, per bin, a tiny bit from what
one might expect analytically from truly iid samples.  One certainly
isn't going to see this variation in the variance from two runs -- the
expected statistical noise would overwhelm the signal as this is a
second moment/cumulant effect, a susceptibility if you like, and hence
VERY difficult to resolve.

What we in fact must do is (I'm pretty sure) compute c[][] analytically,
(which I have done) and then simulate the SAME c[][] directly from the
data.  Then difference them, which should make the components of
c_exp[][] - c_exact[][] presented as vector a zero mean chisq
distributed object with N! - (N-1)! DoF or something like that -- 120-24
= 96 for N = 5?  Here I'll defer to experts, eventually.