THIS IS A WORKING STS SERIAL as far as I can tell.  It correctly reveals known weak generators.  It seems to pass known "decent" generators.  It is very sensitive to certain kinds of failure.
I now should redo rgb_bitdist to be the NON-overlapping version of this
test, at the very least comparing the freq output (generated with the
same code, I think, ported the other way) to a multinomial chisq on the
expected values.  If I can actually check the distribution of each
number against an asymptotically normal curve, that would be great too.
If I could IN THE SAME TEST check the asymptotic fraction -- something
like the monkey tests do -- that would be just fabulous!  Especially if
I use a normal and make the comparison very good and not empirical.