This checks in the ChangeLog itself, as well as the last few minor changes in the tests run with the -a command and the test listing function in the CLI.  I think I declare this to be:
    2.27.11

and we'll start working on 12.  I think this actually fixes
all the outstanding bugs except for the problem in sums and
the problem in operm5 in diehard (and of course there are tests
half implemented as well).

I THOUGHT I'd fixed diehard_sums, actually, but it still most definitely
fails on large -p.  I'll have to check sums out explicitly to see why it
doesn't seem to work at the moment.  It's possible I dropped the fixes
when recently changing systems.

Anyway, here it is.  Bring on the bugs.