This actually checks in so that we can pop the snapshot number in just a minute -- a major bugfix relative to Rdh.  The two problems addressed herein are:
 a) In order to be able to run many tests, one after another, on many
rngs, one after another, I have to be more careful than I have been
about allocating and freeing test resources on the one hand, rngs (which
must be freed) on the other hand, and resetting the static bit buffers
used in bits.c on the third hand.

 b) I also needed to set up startup with a split into two sections, one
that runs only one time period and one that runs every time a new test
is created, executed, and destroyed (with everything reset at the end to
run a new test).

The good thing beyond Rdh in these changes is that several of them are
equally necessary in a GUI version e.g. gdieharder.  So nothing being
done here is wasted...