This is a 2.27.14 checkin.  I'll shoot it off to Dirk who is waiting on it.  I noted, however, that my NEW permutations test FAILS (or should fail) mt19937!  It produces too GOOD a spread of permutations, consistently!  This is fascinating information.  It means that perhaps operm5 is NOT broken; maybe getting permutations to work out multinomially is the most difficult test one can put an rng to.
Something to look at later.

This checkin should fix dieharder and libdieharder so that one can run
multiple tests (including invocation of the same test or different rngs)
from a single dieharder call.  This cannot happen in dieharder -- don't
worry.  But it can and will happen in Rdh and in gdieharder, so it is an
important fix nonetheless.