OK, this is a checkin of dieharder 3.28.0beta.  It is NOT fully decrufted, but seems to mostly nearly hopefully all work.  That probably means that there are only a dozen or two bugs.  There are also a few API features I haven't implemented in the UI yet -- specifically the reporting of errors (like a rewind of a file in mid-test).  So this ISN'T really a beta -- more like an alpha.
Dirk, please do not start converting this over into Rdh yet. I'm
checking it in for two reasons -- one is that I have to remove a whole
pile of files to decruft and svn won't let me until I check in.  Another
is that I NEED to checkin -- it makes me nervous to have this large a
delta not checked in.  There are probably a couple of critical sources
I've forgotten to add entirely and I won't know until I check in and
check out and build a fresh clean copy.

MOST of what I've done, from Rdh's point of view, should be invisible
after Rdh is (fairly minimally) hacked one last time.  Basically Rdh
should use its own version of set_globals.c (or patch mine, or ifdef
mine).  Note that there are a lot fewer variables, and this list may
shrink a bit more.

ntuple's meaning hasn't changed, and you already handle that.

Seed and strategy work together -- the latter is a new variable and
SHOULDN'T affect Rdh, but just in case, here's what it does.

The default strategy for dieharder is to reseed once when a rng is
chosen.  In the default output view, the seed is written to the rng
information part of the header.  That way if one wishes to reproduce
a test result, one can enter the seed with -S seed.

However, this is actually a PROBLEM if one runs multiple tests from this
one seed.  If one runs tests out of order, the results will be
different.  This is true for me running all the tests in order via
dieharder -a if I should ever change test order, and is true for Rdh if
one runs first one test, then another in different orders, from the same
single specification of rng and/or Seed.

Also, there may be situations where one wants to run a single test
multiple times, each time from a newly selected seed to (in essence)
determine if some seeds are "bad" for a given rng.  dieharder doesn't
yet support that, but I think that in R it would be pretty easy.

Setting strategy to anything nonzero (say, 1) causes dieharder to reseed
the random number generator at the beginning of any test.  If -S seed
was NOT specified, it just generates a new random seed, so that one
could run e.g. diehard_birthdays 100 times in a loop and each one would
reseed anew with a new random seed.  If -S seed IS specified, it uses
the specified seed.  If a file is being used for input (not stdin) it
forces a rewind at the beginning of each test, which is actually not a
bad thing to do as it conserves rands.  (I hope, I haven't yet tested
this latter feature much yet but it should work.:-).

SO, Rdh will probably just leave strategy = 0 alone and either set the
global value of Seed for one-time initialization with a fixed seed or
not, accepting a one-time random seed.  But you CAN support strategy if
you ever think you need to.

(From my point of view its primary purpose is to make the creation of a
validation run trivial -- if I run 

  dieharder -a -S 1 -s 1 

I generate a validation table.  If I run 

  dieharder -a -S 1 -s 1 -D test_name -D pvalues

I generate a very sparse validation table (basically, just test name and
pvalue).  You can throw a -c ' ' in there if you want white space
separation or -c ',' if you prefer comma separated values, etc.  You
have nearly complete control over dieharder's output at this point, see:

  dieharder -F 

for a listing of output control flags.  dieharder -l and dieharder -h
and dieharder -g -1 all work as before, but I've completely flattened
and rationalized test-space so it works just like rng-space (and I mean
JUST like it -- very similar setup.

I completely changed (seriously streamlined and cleaned up) the test
call procedure, so a SINGLE run_test() routine does pretty much all the
work, a SINGLE output() routine does all of THAT work, and so that all
the dieharder CLI-specific stuff is done in parsecl(), and then only if
you enter specific commands, or in dieharder.c (main()).  I tried to
label things that are CLI specific there as well.

If you want to grab a copy of this and build it and play with it, feel
free.  As soon as this checkin is complete, though, I'm going to start
decrufting and checking to be sure I have all the required modules
actually in the repo.