This is a WORKING snap and bump to 3.29.6beta.  I actually fixed several things that I broke before in the rng selection process.  New features:
 rng_kiss -- a damn fine rng.  Faster than mt, better than mt except for
period.

 rng_XOR -- Select this rng, and a list of others, e.g.

   ./dieharder -g 207 -g 208 -g 14 -g 6 -g 205 -a

dieharder will then return the output of 208 (kiss), 14 (mt19937_1999),
6 (gfsr4) and 205 (aes) all xor'd together.  Period infinite, no LESS
random than the MOST random of the generators alone.  The price you pay
is sure, 2, 3, 4 times slower.  But this is now the official gold
standard dieharder testing generator, as finding something randomer will
be difficult and of longer period impossible (what is the least common
multiple of 19937, 121, and all the rest? 2 to that power, like that).

I'm working on superkiss, a vectorized version of kiss with an insanely
long period, but the double precision part is broken and I don't see
why.  The integer part works.  I'll figure it out maybe tonight, and
have a few other Marsaglia generators to add.

Then I'll return (finally) to tests, with the gold standard generator
well and truly in place.  At least three dieharder (diehard) tests are
broken, and I'd like to fix at least ONE of them before I get bogged
down teaching again.

   rgb