This checkin has some very useful documentation in it more than anything else (the rng_superkiss is probably still under development, as always I'm finding myself buried with teaching and unable to quite finish anything I start, sigh...:-(.  GoodPracticeRNG.pdf is a very interesting document, worth reading on general aspects of RNGs and tests.  The KISS code I've implemented is derived/copied from here, with permission.
But the main reason I'm checking in is to save the idea Allyn Shell just
contributed.  He's building ANNs and has observed that -- as is well
known -- RNGs started with adjacent seeds often produce sequences that
are correlated, at least in the short run.  (Initializing an set of ANN
builds with correlated weights IS a short run of calls, and has
undesireable consequences if you're trying to broadly search for an
optimum).

Naturally, a "perfect RNG" according to the null hypothesis would be
perfectly decorrelated for any two seeds, no matter how bitlevel
correlated the seeds are.  MANY generators will not be, because the
iterated maps are discretized chaotic paths and have some sort of
underlying Lyupanov exponent(s) that describes how rapidly to initially
adjacent trajectories diverge.

But this is EXACTLY the kind of thing I'm very interested in measuring
in dieharder, and a very modest clone/hack of the supergenerator is just
what is needed to enable it.  If indeed nearly all rngs (that don't have
a hidden "warmup" period at initialization) exhibit easily detectable
bitlevel/sequential correlations when started with seeds that share all
but one, all but two, all but three... bits, and one can extract a
"sequential correlation time" averaged over all of the possible permuted
adjacent seeds (per test) those times may well prove to be highly
characteristic if not unique to the generator.  It will also be very
interesting to see how "good" generators stack up against this sort of
test.  

So if I can actually drag myself away from all the fun I'm having
grading papers and cleaning my house this weekend, perhaps I'll dig in
and find out.  Ideas/suggestions for implementation welcome.  My initial
impulse is to just make the supergenerator sufficiently controllable
that it can be used for N copies of a single generator, deliberately
seeded with sequential seeds in any bit window on an otherwise identical
mask, and then NOT use a shuffle table but rather return rngs round
robin from the list.  This can be fed into any test, obviously.  When
this "works" run by hand, if there are interesting results being found I
can then add a UI-level way of just "calling a test" that does this sort
of thing automagically.

   rgb